Move ImageBrower Dialog to image_utils.py
